requirements.txt
numpy
pandas
matplotlib
scipy
pykrx
openpyxl
httpx
python-dotenv
